+++ src/sample/Force.java
@@ -13,7 +13,7 @@ public abstract class Force extends Component {
 
     private boolean isAreaSplash;
 
-    public Force(Point2D position , int range,Role role, double hitSpeed, int HP, int damage, Class targetType, boolean isAreaSplash){
+    public Force(Point2D position , double range,Role role, double hitSpeed, int HP, int damage, Class targetType, boolean isAreaSplash){
         super(position,range,role);
         this.hitSpeed=hitSpeed;
         this.HP=HP;
